requirements.txt
PySimpleGUI
Pillow
requests
validators